commit 290e037ede8cb6ed9214469aea9f72f781fc6fea
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sat Dec 10 17:57:43 2011 -0500

    Fix PaX build error on x64

 arch/x86/lib/rwlock_64.S |    1 +
 1 files changed, 1 insertions(+), 0 deletions(-)

commit f470e9ac2d237bab10bdcb057bb35d5e2bbc1fb8
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sat Dec 10 16:30:30 2011 -0500

    Import changes from pax-linux-2.6.32.48-test132.patch

 arch/x86/include/asm/uaccess_64.h |   45 ++++++++++++++++++-------------------
 arch/x86/kernel/entry_64.S        |    7 ++---
 arch/x86/kernel/head_64.S         |   13 +++++++++-
 arch/x86/lib/rwlock_64.S          |   40 ++++++++++++++++++++++++++++++++
 arch/x86/lib/usercopy_64.c        |    4 +-
 fs/partitions/efi.c               |    8 +++---
 include/asm-generic/atomic64.h    |   12 ++++++++++
 tools/gcc/constify_plugin.c       |    6 ++--
 tools/gcc/stackleak_plugin.c      |    6 ++--
 9 files changed, 100 insertions(+), 41 deletions(-)

Signature-tree: 4aff90a9ac57e0e94ebd3256deec0ce3e219d58c

+++ arch/x86/lib/usercopy_64.c
@@ -161,7 +161,7 @@ long strlen_user(const char __user *s)
 }
 EXPORT_SYMBOL(strlen_user);
 
-unsigned long copy_in_user(void __user *to, const void __user *from, unsigned len)
+unsigned long copy_in_user(void __user *to, const void __user *from, unsigned long len)
 {
 	if (access_ok(VERIFY_WRITE, to, len) && access_ok(VERIFY_READ, from, len)) {
 
@@ -184,7 +184,7 @@ EXPORT_SYMBOL(copy_in_user);
  * it is not necessary to optimize tail handling.
  */
 unsigned long
-copy_user_handle_tail(char __user *to, char __user *from, unsigned len, unsigned zerorest)
+copy_user_handle_tail(char __user *to, char __user *from, unsigned long len, unsigned zerorest)
 {
 	char c;
 	unsigned zero_len;
